+++ Engine2/Simulation Runtime/Engine/System/Input/System/SInputHistory.swift
@@ -1,6 +1,6 @@
 /// Projects authoritative input into bounded diagnostic history before cleanup.
 struct SInputHistory: PSystem {
-    mutating func update(world: inout World, deltaTime _: Float) {
+    mutating func update(world: inout World, deltaTime _: Double) {
         world.inputHistory.record(input: world.input)
     }
 }
