+++ Engine2/Simulation Runtime/Engine/System/Position/Protocol/PMovable.swift
@@ -1,18 +1,18 @@
 /// Capability for positioned entity facades backed by translational motion state.
 ///
 /// Accessors expose integrated velocity, persistent acceleration intent, and
-/// the current tick's acceleration and impulse contributions from `CMotion`.
+/// the pending acceleration and impulse contributions from `CMotion`.
 /// They are ergonomic live reads for game code and tooling, not the iteration
 /// surface for movement systems.
 protocol PMovable: PPositionable {
-    var acceleration: SIMD3<Float> { get }
+    var acceleration: SIMD3<Double> { get }
     var accelerationIntent: CMotion.AccelerationIntent { get }
-    var impulse: SIMD3<Float> { get }
-    var velocity: SIMD3<Float> { get }
+    var impulse: SIMD3<Double> { get }
+    var velocity: SIMD3<Double> { get }
 }
 
 extension PMovable {
-    var acceleration: SIMD3<Float> {
+    var acceleration: SIMD3<Double> {
         guard let motion = world.motionComponents[self.id] else {
             fatalError("There is no motion component for the movable entity with ID: \(self.id)")
         }
@@ -26,14 +26,14 @@ extension PMovable {
         return motion.accelerationIntent
     }
 
-    var impulse: SIMD3<Float> {
+    var impulse: SIMD3<Double> {
         guard let motion = world.motionComponents[self.id] else {
             fatalError("There is no motion component for the movable entity with ID: \(self.id)")
         }
         return motion.impulse
     }
 
-    var velocity: SIMD3<Float> {
+    var velocity: SIMD3<Double> {
         guard let motion = world.motionComponents[self.id] else {
             fatalError("There is no motion component for the movable entity with ID: \(self.id)")
         }
